Move sorting related utils to sort.utils.ts
